+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [3.4.0](https://github.com/cedx/ngx-webstorage.js/compare/v3.3.0...v3.4.0)
+- Added support for the [storage events](https://developer.mozilla.org/en-US/docs/Web/API/Window/storage_event).
 
 ## Version [3.3.0](https://github.com/cedx/ngx-webstorage.js/compare/v3.2.0...v3.3.0)
 - Added support for [Angular Universal](https://angular.io/guide/universal).